+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/MiniAppDownloaderSpec.kt
@@ -6,6 +6,7 @@ import com.nhaarman.mockitokotlin2.mock
 import com.rakuten.tech.mobile.miniapp.api.*
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
 import com.rakuten.tech.mobile.miniapp.storage.CachedMiniAppVerifier
+import com.rakuten.tech.mobile.miniapp.storage.ManifestApiCache
 import com.rakuten.tech.mobile.miniapp.storage.MiniAppStatus
 import com.rakuten.tech.mobile.miniapp.storage.MiniAppStorage
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -29,6 +30,7 @@ class MiniAppDownloaderSpec {
     private val storage: MiniAppStorage = mock()
     private val miniAppStatus: MiniAppStatus = mock()
     private val verifier: CachedMiniAppVerifier = mock()
+    private val miniApiCache: ManifestApiCache = mock()
     private lateinit var downloader: MiniAppDownloader
     private val dispatcher = TestCoroutineDispatcher()
     private val testMiniApp = TEST_MA.copy(
@@ -43,6 +45,7 @@ class MiniAppDownloaderSpec {
             initStorage = { storage },
             initStatus = { miniAppStatus },
             initVerifier = { verifier },
+            initManifestApiCache = { miniApiCache },
             coroutineDispatcher = dispatcher
         )
         downloader.updateApiClient(apiClient)